refactor: move commands into separate modules
