+++ src/cmd/publish.rs
@@ -1,0 +1,8 @@
+/// A Redis PUBLISH command.
+#[allow(dead_code)]
+pub struct Publish {
+    channel: String,
+    message: String,
+}
+
+impl Publish {}
