+++ src/cmd/subscribe.rs
@@ -1,0 +1,11 @@
+#[allow(dead_code)]
+pub struct Subscribe {
+    channels: Vec<String>,
+}
+
+impl Subscribe {}
+
+#[allow(dead_code)]
+pub struct Unsubscribe {
+    channels: Vec<String>,
+}
