+++ src/cmd/unsubscribe.rs
@@ -1,0 +1,6 @@
+#[allow(dead_code)]
+pub struct Unsubscribe {
+    channels: Vec<String>,
+}
+
+impl Unsubscribe {}
